chore(dev): improve layer layout, improve Input layout, and fix types
